AI-Forensics-DeepFake-Detection/requirements.txt
Package                       Version
----------------------------- -------------------
albumentations                0.5.2
argon2-cffi                   20.1.0
async-generator               1.10
attrs                         20.3.0
autopep8                      1.5.4
av                            6.2.0
backcall                      0.2.0
backports.functools-lru-cache 1.6.3
black                         19.10b0
bleach                        3.3.0
certifi                       2020.12.5
cffi                          1.14.5
chardet                       4.0.0
click                         7.1.2
cloudpickle                   1.6.0
colorama                      0.4.4
cycler                        0.10.0
cytoolz                       0.11.0
dask                          2021.3.0
decorator                     5.0.6
defusedxml                    0.7.1
efficientnet-pytorch          0.7.0
entrypoints                   0.3
ffmpeg                        1.4
ffmpeg-python                 0.2.0
Flask                         1.1.2
Flask-Cors                    3.0.10
future                        0.18.2
idna                          2.10
imagecodecs-lite              2019.12.3
imageio                       2.9.0
imgaug                        0.4.0
importlib-metadata            3.10.0
ipykernel                     5.5.3
ipython                       7.16.1
ipython-genutils              0.2.0
ipywidgets                    7.6.3
iso8601                       0.1.14
itsdangerous                  1.1.0
jedi                          0.17.2
Jinja2                        2.11.3
joblib                        1.0.1
jsonschema                    3.2.0
jupyter                       1.0.0
jupyter-client                6.1.12
jupyter-console               6.4.0
jupyter-core                  4.7.1
jupyterlab-pygments           0.1.2
jupyterlab-widgets            1.0.0
kiwisolver                    1.3.1
m3u8                          0.8.0
MarkupSafe                    1.1.1
matplotlib                    3.3.4
mistune                       0.8.4
nbclient                      0.5.3
nbconvert                     6.0.7
nbformat                      5.1.3
nest-asyncio                  1.5.1
networkx                      2.5
notebook                      6.3.0
numpy                         1.19.5
olefile                       0.46
packaging                     20.9
pandas                        1.1.5
pandocfilters                 1.4.2
parso                         0.7.1
pathspec                      0.8.0
pickleshare                   0.7.5
Pillow                        8.1.2
pip                           21.0.1
prometheus-client             0.10.0
prompt-toolkit                3.0.18
protobuf                      3.15.7
pycodestyle                   2.6.0
pycparser                     2.20
Pygments                      2.8.1
pyparsing                     2.4.7
PyQt5                         5.12.3
PyQt5-sip                     4.19.18
PyQtChart                     5.12
PyQtWebEngine                 5.12.1
pyrsistent                    0.17.3
python-dateutil               2.8.1
pytz                          2021.1
PyWavelets                    1.1.1
pywin32                       300
pywinpty                      0.5.7
PyYAML                        5.4.1
pyzmq                         22.0.3
qtconsole                     5.0.3
QtPy                          1.9.0
regex                         2020.6.8
requests                      2.25.1
scikit-image                  0.17.2
scikit-learn                  0.24.1
scipy                         1.5.3
Send2Trash                    1.5.0
setuptools                    49.6.0.post20210108
Shapely                       1.7.1
six                           1.15.0
tensorboardX                  2.0
terminado                     0.9.4
testpath                      0.4.4
threadpoolctl                 2.1.0
tifffile                      2020.6.3
toml                          0.10.1
toolz                         0.11.1
torch                         1.4.0
torchvision                   0.5.0
tornado                       6.1
tqdm                          4.60.0
traitlets                     4.3.3
typed-ast                     1.4.1
typing-extensions             3.7.4.3
urllib3                       1.26.4
wcwidth                       0.2.5
webencodings                  0.5.1
Werkzeug                      1.0.1
wheel                         0.36.2
widgetsnbextension            3.5.1
wincertstore                  0.2
zipp                          3.4.1
